added 'About Purchases' for shrutibox
Most of the *About Purchases* page contents are same for all apps. We've
already extracted the sections and using them with substitution for
Sadhakam. This time we are using the same parts for Shrutibox too.

* also extracted premium troubleshooting steps. Premium feature's name
and steps to toggle on premium feature should be added to each app's
yaml front matter
* refactored shruti's about purchases page to use the same extracted
sections. Shruti's was a html page, now markdown
* TODO add `image/shrutibox_seal.svg`. (currently displays the alt text)