about purchases: added link that shows how to find upi id
* this link is to a 3rd party blog post filled with ads. But as far as
the content and writing is concerned, this is the best among what we
found. We ensured that we skip to the meat of the content which somewhat
mitigates the confusion that ads could cause.
* we shall verify this content from time to time to see if this is kept
up-to-date with the UPI apps. If we find a better write up, we could
replace it. We could also consider coming up with our own write up as
part of FAQ.